pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "enhance-this"
version = "0.1.0"
authors = [
  { name="Hariharen", email="thisishariharen@gmail.com" },
]
description = "A CLI tool to enhance prompts using Ollama AI."
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "click>=8.0.0",
    "requests>=2.25.0",
    "pyyaml>=6.0",
    "pyperclip>=1.8.0",
    "rich>=10.0.0",
    "importlib-resources; python_version < '3.9'",
]

[project.urls]
"Homepage" = "https://github.com/hariharen9/enhance-this"
"Bug Tracker" = "https://github.com/hariharen9/enhance-this/issues"

[project.scripts]
enhance = "enhance_this.cli:main"

[tool.setuptools.packages]
find = {where = ["."], include = ["enhance_this*"]}
